Move tethys_store.hpp to tethys_store_builder.hpp
